fix: examples中的文件流时长精度问题

+++ examples/qt_gui/src/SimplePlayer.cpp
@@ -79,7 +79,7 @@ void SimplePlayer::onStopRecordBtnClicked()
     ui->player->stopRecoding();
 }
 
-void SimplePlayer::onTotalTimeRecved(int totalTime)
+void SimplePlayer::onTotalTimeRecved(int64_t totalTime)
 {
     totalTime_ = totalTime;
     ui->endTimeLabel->setText(QString::number(totalTime));
